[feature] Kommander Thanos API Test (#38)
* Update kubeaddons

* Add kommander thanos test

* Update to latest kubeaddons v0.10.2

* Add thanos-checker resource definitions

* Cleanup tests

The tests here were getting messy, this separates things into
logical groupings of functionality, including a file specifically
for testing kommander, and one for testing thanos.

Also marked some functionality that belongs upstream.

* Update test/unhandled_addons_test.go
