+++ test/utils_test.go
@@ -1,0 +1,13 @@
+package test
+
+import (
+	"os"
+	"os/exec"
+)
+
+func kubectl(args ...string) error {
+	cmd := exec.Command("kubectl", args...)
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	return cmd.Run()
+}
